+++ Source/FileSystem/FileComparer.cs
@@ -1,0 +1,11 @@
+﻿using System.IO.Abstractions;
+
+namespace FileSystem;
+
+public class FileComparer : IFileComparer
+{
+    public bool AreEqual(IFileInfo info, IFileInfo fileInfo)
+    {
+        throw new NotImplementedException();
+    }
+}
