+++ Source/FileSystem/FileDiffStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace FileSystem;
+
+public enum FileDiffStatus
+{
+    Invalid = 0,
+    Deleted,
+    Unchanged,
+    Modified,
+    Created
+}
