+++ Source/FileSystem/IFileComparer.cs
@@ -1,0 +1,8 @@
+﻿using System.IO.Abstractions;
+
+namespace FileSystem;
+
+public interface IFileComparer
+{
+    bool AreEqual(IFileInfo info, IFileInfo fileInfo);
+}
