+++ Source/FileSystem/IFileSystemPathTranslator.cs
@@ -1,0 +1,8 @@
+﻿using System.IO.Abstractions;
+
+namespace FileSystem;
+
+public interface IFileSystemPathTranslator
+{
+    public string Translate(IFileSystemInfo fileSystemInfo, IFileSystemInfo origin, IDirectoryInfo destination);
+}
